Rename network.py to ofa_xception.py.
