+++ nnabla_nas/contrib/classification/ofa_xception/__init__.py
@@ -1,5 +1,5 @@
-from .network import SearchNet
-from .network import TrainNet
+from .ofa_xception import SearchNet
+from .ofa_xception import TrainNet
 
 
 __all__ = ['TrainNet', 'SearchNet']
